pyproject.toml
[project]
name = "kunskaps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "motor>=3.5.1",
    "openai-whisper>=20240930",
    "pre-commit>=4.0.1",
    "python-multipart>=0.0.19",
    "ruff>=0.8.2"
]

[tool.ruff]
# Set the maximum line length to 79.
line-length = 79

[tool.ruff.lint]
select = ["E", "F", "I", "N", "FAST", "DOC", "PL"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
